% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{editor_text}
\alias{editor_text}
\title{Update HTML Editor on the client}
\usage{
editor_text(
  selector,
  asis = FALSE,
  inputId,
  as_fn = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{selector}{\code{chr} The inputId used for \code{editor_ui}. This is namespaced and \verb{#} is prepended unless \code{asis = TRUE} in which case this is taken verbatim as the selector.}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{inputId}{\code{chr} The inputId to save the content of the TinyMCE editor to. This is namespaced unless \code{asis = TRUE} in which case this is taken verbatim as the inputId.}

\item{session}{A shiny session object.}

\item{text}{Prepopulated Text to be shown in editor.}
}
\value{
\code{NULL} Updates the text in the TinyMCE Editor

\code{NULL/fun} If \code{as_fn = FALSE} \emph{Default} the \code{input[[inputId]]} is updated with the generated HTML.
}
\description{
It allows to update HTML Editor in server side.
}
\examples{
library(shiny)
library(OpenMCE)
# UI
ui <- fluidPage(

  # Setup
  use_editor(),
  titlePanel("HTML Generator"),

  # Text Input 1
  fluidRow(
    column(
      width = 6,
      editor_ui('textcontent'),
      br(),
      actionButton(
        "generatehtml",
        "Generate HTML Code",
        icon = icon("code"),
        class = "btn-primary"
      ), actionButton("updatedata", "Update Editor", icon = icon("edit"))),

    column(
      width = 6,
      uiOutput("rawText")
    )
  )

)

# Server
server <- function(input, output, session) {

  # Generate HTML
  content <- reactiveVal()
  editor_server("textcontent", inputId = "mytext")
  output$rawText <- renderUI({
    req(content())
    shiny::HTML(content())
  })
  observeEvent(input$generatehtml, {
    content(input$mytext)
  })

  observeEvent(input$updatedata, {
    editor_update(selector = "textcontent",
                  text = "<b>Sample Text</b>")

  })

}

# Run App
shinyApp(ui = ui, server = server)
Extract HTML generated

It extracts HTML generated based on text written in editor.

devtools::load_all()

# UI
ui <- fluidPage(# Setup
  use_editor(),
  titlePanel("HTML Generator"),

  # Text Input 1
  fluidRow(
    column(
      width = 6,
      editor_ui('textcontent'),
      br(),
      actionButton(
        "generatehtml",
        "Generate HTML Code",
        icon = icon("code"),
        class = "btn-primary"
      )
    ),

    column(width = 6,
           uiOutput("rawText"))
  ))

# Server
server <- function(input, output, session) {
  # Generate HTML
  content <- reactiveVal()
  editor_server("textcontent")

  observeEvent(input$generatehtml, {
    editor_text("textcontent", inputId = "mytext")
    output$rawText <- renderUI({
      req(input$mytext)
      shiny::HTML(input$mytext)
    })
  })

}

# Run App
shinyApp(ui = ui, server = server)
}
