% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{editor_update}
\alias{editor_update}
\title{Update the TinyMCE instance}
\usage{
editor_update(
  inputId,
  text = "Sample Text",
  session = shiny::getDefaultReactiveDomain(),
  asis = FALSE
)
}
\arguments{
\item{inputId}{\code{chr} inputId of the editor}

\item{text}{\code{chr/shiny.tag/shiny.tag.list} HTML and Shiny tags also supported}

\item{session}{A shiny session object.}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}
}
\value{
Updates the input
}
\description{
Update the TinyMCE instance
}
