% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_openmce.R
\name{tiny_mce_dependency}
\alias{tiny_mce_dependency}
\title{Include the tinyMCE javascript dependencies}
\usage{
tiny_mce_dependency(API = "")
}
\value{
\code{html_dependency}
}
\description{
Include the tinyMCE javascript dependencies
}
