% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_openmce.R
\name{use_editor}
\alias{use_editor}
\title{Loads libraries for this package}
\usage{
use_editor(API = Sys.getenv("TINYMCE_KEY"), use_api = FALSE)
}
\arguments{
\item{API}{\code{lgl} If you wish to specifically use the API, set this to \code{TRUE}}
}
\value{
JS Files
}
\description{
Set up for the package. Required in UI
}
